Updating csv options, replacing column indices type

+++ engine/src/io/Schema.cpp
@@ -130,7 +130,7 @@ Schema Schema::fileSchema(size_t current_file_index) const {
 	return schema;
 }
 
-std::unique_ptr<ral::frame::BlazingTable> Schema::makeEmptyBlazingTable(const std::vector<size_t> & column_indices) const {
+std::unique_ptr<ral::frame::BlazingTable> Schema::makeEmptyBlazingTable(const std::vector<int> & column_indices) const {
 	std::vector<std::string> select_names(column_indices.size());
 	std::vector<cudf::type_id> select_types(column_indices.size());
 	if (column_indices.empty()) {
